backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.1

# Database
sqlalchemy==2.0.25
pymysql==1.1.0
alembic==1.13.1

# Data Processing
pandas==2.2.0
numpy==1.26.3

# HTTP Requests
requests==2.31.0
aiohttp==3.9.3

# Scheduler
apscheduler==3.10.4

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Logging
python-json-logger==2.0.4

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# API Rate Limiting
slowapi==0.1.9

# Prometheus Metrics
prometheus-fastapi-instrumentator==7.1.0

# File Upload
python-multipart==0.0.6

# FMZ API
# FMZ使用官方API，这里保持简洁

# Data Source
akshare>=1.12.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
